refactor: extract Map entity to domain layer
- Move Map struct and all implementation from entities.rs and map.rs to src/domain/entities/map.rs
- Consolidate MapConstants, MapError, MapResult in the domain entity
- Update map.rs to simple re-export for backward compatibility
- Fix test imports to resolve conflicts with multiple new() methods
- Follow clean architecture with domain entity consolidation
